feature/UPSE-103 functional version migrated to new httpClient builder architecture
