+++ pom.xml
@@ -170,7 +170,7 @@
             <!-- for cernunnos -->
                 <groupId>javax.script</groupId>
                 <artifactId>groovy-engine</artifactId>
-                <version>1.1</version>
+                <version>1.0</version>
                 <scope>runtime</scope>
             </dependency>
             <dependency>
@@ -547,6 +547,10 @@
         <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>jcl-over-slf4j</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>javax.script</groupId>
+            <artifactId>groovy-engine</artifactId>
         </dependency>
         <dependency>
             <groupId>ch.qos.logback</groupId>
