+++ src/main/resources/configuration.properties
@@ -18,3 +18,6 @@
 #
 
 userinfo.userid.key=username
+
+connectionTimeoutInMillis=1000
+socketTimeoutInMillis=1000
